Isole le code dans un composant
